Update readStackTif.m
fix uint8 error. if there are too many labels is not a good option

+++ lib/readStackTif.m
@@ -1,7 +1,7 @@
 function [loadedImage, infoImage] = readStackTif(fileName)
 
     infoImage = imfinfo(fileName);
-    loadedImage = uint8(zeros(infoImage(1).Height,infoImage(1).Width,size(infoImage,1)));
+    loadedImage = uint16(zeros(infoImage(1).Height,infoImage(1).Width,size(infoImage,1)));
     for nZ = 1:size(infoImage,1)
         loadedImage(:,:,nZ) = imread(fileName,nZ);
     end
